# docs/specs/swagger.yaml
swagger: "2.0"
info:
  description: "This component is in charge of monitoring and solving the inconsistencies generated by the loss of notifications about purchasing events. See the full project here [fury_credit-card-purchases-consist](https://github.com/mercadolibre/fury_credit-card-purchases-consist)"
  version: "0.0.0"
  title: "CryptoBot Messaging API"
paths:

  /messages/update:
    post:
      summary: "Exposed endpoint used to push new messages sent by the user"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: "body"
          name: "UpdateMessage"
          required: true
          schema:
            $ref: "#/definitions/UpdateMessageSchema"
      responses:
        '200':
          description: OK

  /messages/news/send:
    post:
      summary: "Exposed endpoint used to send crypto news messages to a user"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: "body"
          name: "NewsMessage"
          required: true
          schema:
            $ref: "#/definitions/NewsMessage"
      responses:
        '200':
          description: OK
        '500':
          description: Internal server error
          schema:
            $ref: "#/definitions/InternalServerErrorResponse"
definitions:
  UpdateMessageSchema:
    type: "object"
    properties:
      update_id:
        type: "integer"
        format: "int64"
        example: 1351
      message:
        $ref: "#/definitions/Message"

  Message:
    type: "object"
    properties:
      message_id:
        type: "integer"
        format: "int 64"
        example: 23965
      chat:
        $ref: "#/definitions/Chat"
      text:
        type: "string"
        format: "string"
        example: "Hi crypto bot!"

  Chat:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 67479
      type:
        type: "string"
        format: "string"
        example: "Private"

  NewsMessage:
    type: "object"
    properties:
      chat_id:
        type: "integer"
        format: "int64"
        example: 145688
      title:
        type: "string"
        format: "string"
        description: "news title"
        example: "Ethereum podría destronar al bitcoin en valor de mercado con su última actualización"
      link:
        type: "string"
        format: "string"
        example: "https://www.eleconomista.es/mercados-cotizaciones/noticias/11684467/03/22/Ethereum-podria-destronar-al-bitcoin-en-valor-de-mercado-con-su-ultima-actualizacion.html"

  InternalServerErrorResponse:
    type: "object"
    properties:
      status:
        type: "integer"
        format: "int32"
        example: 500
      message:
        type: "string"
        example: "internal server error"
      error:
        type: "string"
        example: "internal_server_error"
      cause:
        type: "array"
        items:
          type: "string"